chore - cleans up a missing - moves misc not page components out of pages directory
