+++ src/components/ui/card.tsx
@@ -101,4 +101,5 @@ export {
   CardTitle,
   CardDescription,
   CardContent,
+  CardBody,
 };
